refactored subscriber to postpone topic type analysis until first message received
Using the rostopic approach to subscribe to a topic will fail if the
topic is not yet published. It also did ignore remapping of topic names.
Now, we subscribe to a topic, listening to rospy.msg.AnyMsg and retrieve
the actual topic type on first received message and then trigger
deserialization manually.